pyproject.toml
[tool.poetry]
name = "llm_serv"
version = "2025.9.8"
description = "LLM Service"
authors = ["Stefan Dumitrescu <dumitrescu.stefan@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
pydantic = "^2.11.4"
pytest = "8.3.3"
fastapi = "^0.115.0"
uvicorn = "^0.32.1"
PyYAML = "^6.0.1"
rich = "^13.9.4"
pillow = "10.4.0"
dateparser = "^1.2.1"
pytest-asyncio = "^0.26.0"
requests = "^2.32.4"
openai = "^1.97.0"
aioboto3 = "^14.3.0"
colorama = "^0.4.6"
google-genai = "^1.27.0"
python-box = "^7.3.2"

[tool.pdm.dev-dependencies]
test = [
    "pytest>=7.4.4",
    "requests-mock>=1.11.0",
]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
# Enable the flake8-bugbear rules in addition to the defaults.
select = ["E", "F", "B"]
ignore = []
line-length = 140

# Allow unused variables in `__init__.py` files.
[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]
